Create package for handling 3pid processes and move invite processing there
